Implemented the Context class and completed some more tests.
Created the Context class with the snapshot functionaly.
Tested both Context and Context.Snapshot classes.
Reflected the Snapshot implementation to the Actions.
Fixed bugs in ActionGroup and fully tested it.
Implemented the Constraint/ConstraintGroup class and tested them.
Created the Effect class and tested it.
Created a functional interface VariableSupplier which is responsible for the supply of a variable from the context.
Removed .idea/misc.xml due to conflicts problems.
Removed .iml from .gitignore.
Added .idea/misc.xml to .gitignore.